fix int to uint16 cast for createPacket

+++ cmd/rdpgw/protocol/process.go
@@ -188,7 +188,7 @@ func (p *Processor) Process(ctx context.Context) error {
 				binary.Write(buf, binary.LittleEndian, uint32(0))
 				binary.Write(buf, binary.LittleEndian, uint16(HTTP_EXTENDED_AUTH_PAA))
 				binary.Write(buf, binary.LittleEndian, uint16(0))
-				msg := createPacket(message.packetType, buf.Bytes())
+				msg := createPacket(uint16(message.packetType), buf.Bytes())
 				p.tunnel.Write(msg)
 			default:
 				log.Printf("Unknown packet type=0x%x (size %d): %x", message.packetType, message.length, message.msg)
